pyproject.toml
[tool.poetry]
name = "OrcaDemo"
version = "0.1.0"
description = "demos for orca"
authors = ["Orca DB Inc."]
readme = "README.md"
packages = [{ include = "util" }]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
google-cloud-core = "2.3.3"
google-cloud-storage = "2.10.0"
numpy = "1.25.2"
pandas = "2.1.1"
orcalib = "^0.0.60"
accelerate = "^0.28.0"
nltk = "^3.8.1"
tqdm = "^4.66.1"
transformers = "^4.38"
torch = "^2.1.2"
gradio = "^4.14.0"
ipykernel = "^6.25.2"
ipywidgets = "^8.1.1"
ray = "^2.8.0"
torchvision = "^0.16.1"
scikit-learn = "^1.3.2"
seaborn = "^0.13.0"
safetensors = "^0.4"
trl = "^0.8.0"
orca-common = "0.1.3"
feedparser = "^6.0.11"
datasets = ">=2.14.6,<2.15.0"
ragas = "^0.1.7"
fire = "^0.6.0"
pre-commit = "^3.7.0"
llama-index = "<0.10.26"
llama-index-llms-huggingface = "^0.1.4"
wikipedia = "^1.4.0"
llama-index-readers-wikipedia = "^0.1.3"
crewai = "^0.27.2"
python-dotenv = "^1.0.0"
ollama = "^0.1.9"
dspy-ai = "^2.4.9"
spacy = "^3.7.2"
dateparser = "^1.2.0"
stanfordnlp = "^0.2.0"
stanza = "^1.7.0"
peft = "^0.7.1"
sentence-transformers=">0.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"
